Refactor vech
